Feature/update to 420 latest (#23)
Merge Trino version 420 changes.
Update Iceberg Hadoop catalog to work with code refactoring in trino 420,
See Trino release notes for details: https://trino.io/docs/current/release/release-420.html

+++ core/trino-main/src/test/java/io/trino/sql/planner/TestDeleteAndInsertMergeProcessor.java
@@ -76,11 +76,11 @@ public class TestDeleteAndInsertMergeProcessor
         assertThat(outputPage.getPositionCount()).isEqualTo(1);
 
         // The single operation is a delete
-        assertThat(TINYINT.getLong(outputPage.getBlock(3), 0)).isEqualTo(DELETE_OPERATION_NUMBER);
+        assertThat((int) TINYINT.getByte(outputPage.getBlock(3), 0)).isEqualTo(DELETE_OPERATION_NUMBER);
 
         // Show that the row to be deleted is rowId 0, e.g. ('Dave', 11, 'Devon')
         Block rowIdRow = outputPage.getBlock(4).getObject(0, Block.class);
-        assertThat(INTEGER.getLong(rowIdRow, 1)).isEqualTo(0);
+        assertThat(INTEGER.getInt(rowIdRow, 1)).isEqualTo(0);
     }
 
     @Test
